+++ src/aux-server/aux-web/aux-player/scene/PlayerGame.ts
@@ -458,6 +458,8 @@ export class PlayerGame extends Game {
                     });
                 } else if (e.type === 'import_aux') {
                     this.importAUX(sim, e.url);
+                } else if (e.type === 'play_sound_url') {
+                    this.playAudio(e.url);
                 }
             })
         );
@@ -1001,8 +1003,20 @@ export class PlayerGame extends Game {
 
             mainControls.controls.minPanX = this.getPanMinX();
             mainControls.controls.maxPanX = this.getPanMaxX();
-            mainControls.controls.minPanY = this.getPanMinY() * -1;
-            mainControls.controls.maxPanY = this.getPanMaxY() * -1;
+
+            mainControls.controls.minPanY = this.getPanMinY();
+
+            if (this.getPanMinY() != null) {
+                mainControls.controls.minPanY = this.getPanMinY() * -1;
+            } else {
+                mainControls.controls.minPanY = null;
+            }
+
+            if (this.getPanMaxY() != null) {
+                mainControls.controls.maxPanY = this.getPanMaxY() * -1;
+            } else {
+                mainControls.controls.maxPanY = null;
+            }
         }
 
         if (!this.getInventoryResizable() || !this.invVisibleCurrent) {